Uses byte instead of int for msg type.

+++ src/main/java/com/mytechia/commons/framework/simplemessageprotocol/MessageFactory.java
@@ -32,7 +32,7 @@ import java.util.Map;
  */
 public class MessageFactory {
 
-    private Map<Integer, IMessageBuilder> messageBuilders = new HashMap<>();
+    private Map<Byte, IMessageBuilder> messageBuilders = new HashMap<>();
 
 
     public void registerMessageBuilder(IMessageBuilder messageBuilder) {
